requirements.txt
aiofiles==22.1.0
aiosqlite==0.18.0
altair==5.1.2
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==23.1.0
Babel==2.11.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==4.1.0
blinker==1.7.0
Brotli==1.0.9
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
comm==0.1.2
cryptography==41.0.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.0.4
executing==0.8.3
extra-streamlit-components==0.1.60
fastjsonschema==2.16.2
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.25.0
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.6
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==7.4.9
jupyter-console==6.6.3
jupyter_core==5.5.0
jupyter-events==0.8.0
jupyter-server==1.23.4
jupyter_server_fileid==0.9.0
jupyter_server_ydoc==0.8.0
jupyter-ydoc==0.2.4
jupyterlab==3.6.3
jupyterlab-pygments==0.1.2
jupyterlab_server==2.22.0
jupyterlab-widgets==3.0.5
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
nbclassic==0.5.5
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.26.1
packaging==23.1
pandas==2.1.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pip==23.3
platformdirs==3.10.0
plotly==5.18.0
plotly-express==0.4.1
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==4.25.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.8.0
pymssql==2.2.10
pyOpenSSL==23.2.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==23.2.0
qtconsole==5.4.2
QtPy==2.2.0
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.6
scipy==1.11.3
Send2Trash==1.8.2
setuptools==68.0.0
sip==6.7.12
six==1.16.0
smmap==5.0.1
sniffio==1.2.0
soupsieve==2.5
stack-data==0.2.0
statsmodels==0.14.0
streamlit==1.28.1
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6
tenacity==8.2.3
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
traitlets==5.7.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.41.2
widgetsnbextension==4.0.5
y-py==0.5.9
ypy-websocket==0.8.2
zipp==3.11.0